HT_st7735: port library to espidf
The st7735 is a TFT driver. The original driver [1] requires the
Arduino Framework, therefore it is not compatible with espif. This
commit ports the library to esp-idf. In addition, it cleans up the
original and creates a wrapper for the SPI drivers. As the SPI driver
might need to be reused in the future, spoiler: LORA, it might be
refactored and moved to a separate library.

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>