gnss: Add gnss support and example
Create a new gnss task that prints the current location on the screen.
Important the app_main task, must not be terminated otherwise the
display goes out of scope. In addition, we can't use an infinite busy
waiting period due to a) power and b) it causes the watchdog timer to
fire and pollute our logs (and probably messes up with the scheduler)

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>