HT_st7735: add {hold, unhold}_pins method
When the device is put into a sleep state, we might want to keep the
display on. In order to do so, we will need to keep deliver power to
display pins. The unhold method on the other hand will let the display
shut off. On USB the displaly consumes ~40mA of current with the Backlit
on.

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>