BatteryMonitor: Introduce Battery monitoring
It will be very important to know how much juice we have left while on
the field. For this reason, we use the ADC1/CHA0 that it is connected to
the Battery to measure the battery voltage and by using a quick linear
assumption of voltage discharge curve, we can output the battery
percentage on the screen. A more accurate interpolation, would require
too costly math operations that are really not necessary.

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>