HT_st7735: add mutex locking
The design of the display driver is just bad. There can be cases, where
even without true parallelism we might have concurrency, because of task
preemption. As a result, we will need to introduce a synchronisation
primitive and here we have a mutex. RTOS mutexes employ priority
inheritance. This means that if a high priority task blocks while
attempting to obtain a mutex (token) that is currently held by a lower
priority task, then the priority of the task holding the token is
temporarily raised to that of the blocking task.

In the fuure we should redesign the display driver or migrate to lgvl
and instead have a task that refreshes the screen every x amount of ms
and the various tasks write to a buffer and not directly to the screen.

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>