enable Power Management
FreeRTOS and esp-idf support power management [1]. In order to enable we
need to change the following configurations:
```
CONFIG_PM_ENABLE=y
CONFIG_FREERTOS_USE_TICKLESS_IDLE=y
CONFIG_FREERTOS_HZ=1000
```

This is normally done via menuconfig, but as platofrmio is
auto-generating this file, commiting the sdkconfig is not a good idea.
For this reason, we add them in a separate file and introduce an extra
script, which overwrites the configuration options.

Now the main configuration is happening via the `esp_pm_configure` and
this commits sets the maximum and minimum clock frequencies as well as
letting the SoC to go to light sleep mode [2]. During the light sleep
mode a lot of peripherals are reset, including GPIO, so a modification
is required to hold them during the sleep phase. In order for the system
to actually go to sleep, all the tasks need to be idling for at least
`CONFIG_FREERTOS_IDLE_TIME_BEFORE_SLEEP` x `1/CONFIG_FREERTOS_HZ` sec.

Moreover PM has the concept of **Power Management Locks**.
Applications have the ability to acquire/release locks in order to
control the power management algorithm. When an application acquires
a lock, the power management algorithm operation is restricted, ie the
clock frequency will not drop below the set value. This might or might
not be necessary if the Dynamic Frequency Scaling is not enabled as is
the case now. In the future we should reconsider using DFS, but it seems
a bit scary. We an always optimise later, but for let's not make our
lives harder than what they need to be.

[1] https://docs.espressif.com/projects/esp-idf/en/stable/esp32/api-reference/system/power_management.html
[2] https://docs.espressif.com/projects/esp-idf/en/stable/esp32/api-reference/system/sleep_modes.html

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>