main: Hello World, again...
Oh the irony...
With the complexity of the project rising and the fact the UART
monitoring can mess up with the power settings, we need another way to
ensure that we are not stuck, or the clocks are misbehaving. What better
way than to blink a led. In the release configuration, it will be
removed.

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>