platformio: add monitor filters for debugging
Embeded devices don't have an OS, this means there is no Memory
protection and seg faults. A wrong pointer caused a chain mayhem which
in turn made some mutexes to never be released which then caused the IWD
timeout to fire and crash the processor. The only way this was able to
be debugged was via backtracing. Doing it manually, is a huge pain,
therefore let's enable the exception decoder for the dev builds that we
want now.

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>