astrolavos: bug: fix skipping id in updateDeviceCoordinates
We are currently using the id as the index to update the coordinates.
However this is not right, as the id does not represent the position of
the device in the array. This means that we are updating the wrong
device everytime! fix it by calling directly the method from from the
found device

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>