Astrolavos: introduce Isolation mode
Isolation mode is a state of lower energy, when we are only transmitting
and receiving data, but we are not performing any calculations or using
the display. It is currently edge triggered by pressing the integrated
USR_SWITCH of the heltec wireless board.

Unfortunately, the way FreeRTOS works is that an interrupt cannot wake
up a task that is sleeping via the vTaskDelay. As a result we still have
to wakeup and check up if anything is changed.

Issue: #2

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>