Astrolavos: Refactor device updata to use device_data_t
We were not ambitious enough in the past and we assumed we are only
going to transmit our location. Wrong. In future commits we are going to
support the exchange of more data than just the location. IN order to
accommodate this we need to generalise the payload of the device update
methods.

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>