Astrolavos: Introduce setup mode
The system needs to have a configuration mode. In this mode the user
will be able to setup different parameters. In order to enter this mode
the user will need to press the usr_switch button while the system is
booting. Once the boot process is finished, the setupMode method is
called, the functions associated with this mode are called and in the
end the system is rebooted.

Currently there are no setup functions. These will be added in a future
commit.

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>