lora: Add RadioLib HAL library
We will be using RadioLib[1] for LoRa. In order to make it function with
esp-idf and ESP32S3, we will need to provide a HAL library, that
realises the communication with the physical world. The library is
copied from [2] with the addition of a fix in the SPITransfer, to handle
a NOP command of 0 len.

[1] https://github.com/jgromes/RadioLib
[2] // https://github.com/IanBurwell/DynamicLRS/blob/c26f7f8dcca0c1b70af0aa6aee3aba3a1652aba6/components/DLRS_LoRadio/radiolib_esp32s3_hal.hpp

Signed-off-by: Fotis Xenakis <foxen@windowslive.com>
Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>