lora: Optimise radio parameters for range
We were currently using the default radio parameters for the SX1262.
They are a good middle ground between bandwidth and range, but for our
usecase the maximum theoretical range was 1.8km. We are changing them to
improve the range to 5.8km[1]. This comes at the cost of higher air time
and lower bandwidth, therefore we will need to reduce the size of our
payload, to keep our air time less than a second.

[1] https://www.semtech.com/design-support/lora-calculator

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>