gnss: change TS to uint32_t
The timestamp was now int64_t. This is way too high and we are not
expecting to operate until the end of time. Change it to uint32_t which
still allows for ~50day of continuous operation. By reducing it by 4
bytes, we reduce the LoRa payload by ~20% and decrease our air time to
less than a second

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>