AstrolavosPairedDevice: Overwrite the received ts with local one.
As there is no syncing mechanism in place, the ts of the origin is of
not much use (there is no concept of relaying). Let's overwrite that
with our local one, when we receive a new message. This means that we
can further reduce the payload size. We will followup later.

Issue: #27
Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>