Astrolavos: Add a visual when a device's data is stale
If another device has not transmitted for some time, we still want to
see its last location, but on the other side we also want to to know
that this location might be wrong. We introduce an isStale method that
checks if the last received message was more than 5 minutes ago then it
outputs the device name in red background. If this device had also
indicated that they want to meet then the bg turns Yellow.

Fixes: #6
Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>